requirements.txt
fastapi
uvicorn
pydantic[email]
vyper-config
tenacity
dagster-graphql
alembic
starlette 
psycopg2-binary
sqlalchemy-utils